@@ -9,11 +9,11 @@
     "source": "https://github.com/gubler/collection"
   },
   "require": {
-    "php": "^7.1.3 || ^8.0",
-    "symfony/var-dumper": "^4.1 || ^5.1"
+    "php": "8.*",
+    "symfony/var-dumper": "6.*"
   },
   "require-dev": {
-    "mockery/mockery": "^1.4",
+    "mockery/mockery": "^1.5",
     "phpunit/phpunit": "^9.5"
   },
   "autoload": {
