@@ -2,33 +2,19 @@
 
 namespace Gubler\Collection;
 
-/**
- * Class HigherOrderCollectionProxy
- */
 class HigherOrderCollectionProxy
 {
     /**
      * The collection being operated on.
-     *
-     * @var \Gubler\Collection\Collection
      */
-    protected $collection;
+    protected Collection $collection;
 
     /**
      * The method being proxied.
-     *
-     * @var string
      */
-    protected $method;
+    protected string $method;
 
-    /**
-     * Create a new proxy instance.
-     *
-     * @param  \Illuminate\Support\Collection  $collection
-     * @param  string  $method
-     * @return void
-     */
-    public function __construct(Collection $collection, $method)
+    public function __construct(Collection $collection, string $method)
     {
         $this->method = $method;
         $this->collection = $collection;
@@ -36,11 +22,8 @@ class HigherOrderCollectionProxy
 
     /**
      * Proxy accessing an attribute onto the collection items.
-     *
-     * @param  string  $key
-     * @return mixed
      */
-    public function __get($key)
+    public function __get(string $key): mixed
     {
         return $this->collection->{$this->method}(function ($value) use ($key) {
             return is_array($value) ? $value[$key] : $value->{$key};
@@ -49,12 +32,8 @@ class HigherOrderCollectionProxy
 
     /**
      * Proxy a method call onto the collection items.
-     *
-     * @param  string  $method
-     * @param  array  $parameters
-     * @return mixed
      */
-    public function __call($method, $parameters)
+    public function __call(string $method, array $parameters): mixed
     {
         return $this->collection->{$this->method}(function ($value) use ($method, $parameters) {
             return $value->{$method}(...$parameters);
